+++ src/stack.rs
@@ -1,5 +1,29 @@
 use crate::vm::{RegTypes, VM};
 
+#[derive(Debug)]
+pub struct VMStack {
+    pub stack: Vec<u64>,
+    pub types: Vec<RegTypes>,
+}
+
+impl VMStack {
+    pub fn new(cap: usize) -> VMStack {
+        VMStack {
+            stack: (Vec::with_capacity(cap / 2)),
+            types: (Vec::with_capacity(cap / 2)),
+        }
+    }
+
+    pub fn push(&mut self, dat: u64, dattype: RegTypes) {
+        self.stack.push(dat);
+        self.types.push(dattype);
+    }
+
+    pub fn pop(&mut self) -> (Option<u64>, Option<RegTypes>) {
+        (self.stack.pop(), self.types.pop())
+    }
+}
+
 pub fn op_push(vm: &mut VM) {
     // 0x80, size: 2
     // push Rsrc
@@ -9,8 +33,7 @@ pub fn op_push(vm: &mut VM) {
     let r_type: RegTypes = vm.reg_types[r_src_ind];
 
     // Pushes value, then type.
-    vm.stack.push(val);
-    vm.stack.push((r_type as u64) + 1);
+    vm.stack.push(val, r_type);
 
     vm.ip += 2;
     return;
@@ -20,7 +43,10 @@ pub fn op_pop(vm: &mut VM) {
     // 0x81, size: 2
     // pop Rdest
     let r_dest_ind: usize = vm.memory[(vm.ip + 1)] as usize;
-    let n_type: u64 = match vm.stack.pop() {
+
+    let (val_opt, r_type_opt) = vm.stack.pop();
+
+    let r_type: RegTypes = match r_type_opt {
         Some(val) => val,
         None => {
             panic!(
@@ -29,21 +55,8 @@ pub fn op_pop(vm: &mut VM) {
             );
         }
     };
-    let r_type: RegTypes = match n_type {
-        0x1 => RegTypes::uint64,
-        0x2 => RegTypes::int64,
-        0x3 => RegTypes::float64,
-        0x4 => RegTypes::StrAddr,
-        0x5 => RegTypes::address,
-        other => {
-            panic!(
-                "CRITICAL: Stack member metadata corrupt!\n\tUnknown data type: {}\n\tat IP = {}",
-                other, vm.ip
-            );
-        }
-    };
 
-    let val = match vm.stack.pop() {
+    let val = match val_opt {
         Some(val) => val,
         None => {
             panic!(
@@ -64,8 +77,7 @@ pub fn op_pushall(vm: &mut VM) {
     // 0x82, size: 1
     // pushall - pushes all register values into the stack. (with metadata - types)
     for i in 0..vm.registers.len().saturating_sub(1) {
-        vm.stack.push(vm.registers[i]);
-        vm.stack.push((vm.reg_types[i] as u64) + 1); // pushing metadata next to value
+        vm.stack.push(vm.registers[i], vm.reg_types[i]);
     }
 
     vm.ip += 1;
@@ -76,7 +88,9 @@ pub fn op_popall(vm: &mut VM) {
     // 0x83, size: 1
     // popall - pops last 32 values from stack in registers. Normally used after pushall
     for i in vm.registers.len().saturating_sub(1)..0 {
-        let n_type = match vm.stack.pop() {
+        let (dat_opt, rtype_opt) = vm.stack.pop();
+
+        let r_type: RegTypes = match rtype_opt {
             Some(v) => v,
             None => {
                 panic!(
@@ -85,7 +99,7 @@ pub fn op_popall(vm: &mut VM) {
                 );
             }
         };
-        let val = match vm.stack.pop() {
+        let val = match dat_opt {
             Some(v) => v,
             None => {
                 panic!(
@@ -96,19 +110,7 @@ pub fn op_popall(vm: &mut VM) {
         };
 
         vm.registers[i] = val;
-        vm.reg_types[i] = match n_type {
-            0x1 => RegTypes::uint64,
-            0x2 => RegTypes::int64,
-            0x3 => RegTypes::float64,
-            0x4 => RegTypes::StrAddr,
-            0x5 => RegTypes::address,
-            other => {
-                panic!(
-                    "CRITICAL: Stack member metadata corrupt!\n\tUnknown data type: {}\n\tat IP = {}",
-                    other, vm.ip
-                );
-            }
-        }
+        vm.reg_types[i] = r_type;
     }
 
     vm.ip += 1;